Quick and dirty hack for Bug line 156(surprise me
Reported by Astrid that surprise me went to the main menu almost
immediately. I need to check on the robot if we get a special
cancel event when a surprise thing is closed again. For now we
just reset the active task in the master and wait for the next
command. We leave the MMUI where it is at the moment.

git-svn-id: https://repo.acin.tuwien.ac.at/v4r/trunk/ros/hobbit_hydro_src@10020 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e