new method to make sure that the arm is usable
This method is here to make sure that we can use the arm and the robot can
navigate. If not we ask the user to let the robot put the arm back in and
confirm if there is enough space to move the arm around.
As this needs to be tested though it is not yet included in any scenario.

git-svn-id: https://repo.acin.tuwien.ac.at/v4r/trunk/ros/hobbit_hydro_src@10025 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e